Ultimas modificaciones en los tesst de integración
