+++ app/src/test/java/com/teamtwo/apilol/integrationTestDI.kt
@@ -2,18 +2,11 @@ package com.teamtwo.apilol
 
 import com.example.data.ChampionsLocalDataSource
 import com.example.data.ChampionsRemoteDataSource
-import com.example.data.LocalDataSourceItems
-import com.example.data.RemoteDataSourceItems
 import com.example.domain.Champion
-import com.example.domain.Item
 import com.teamtwo.apilol.di.champions.ChampionsComponent
 import com.teamtwo.apilol.di.champions.ChampionsDataModule
 import com.teamtwo.apilol.di.champions.ChampionsUseCaseModule
-import com.teamtwo.apilol.di.items.ItemsComponent
-import com.teamtwo.apilol.di.items.ItemsDataModule
-import com.teamtwo.apilol.di.items.ItemsUseCaseModule
 import com.teamtwo.apilol.ui.champions.localChampion
-import com.teamtwo.apilol.ui.items.localItem
 import dagger.Component
 import dagger.Module
 import dagger.Provides
@@ -23,9 +16,7 @@ import javax.inject.Singleton
 @Component(modules = [
     TestAppModule::class,
     ChampionsDataModule::class,
-    ChampionsUseCaseModule::class,
-    ItemsDataModule::class,
-    ItemsUseCaseModule::class
+    ChampionsUseCaseModule::class
 ])
 interface TestComponent : ChampionsComponent {
 
@@ -38,16 +29,7 @@ interface TestComponent : ChampionsComponent {
     }
 }
 
-interface ItemTestComponent : ItemsComponent {
 
-    val localDataSource: LocalDataSourceItems
-    val remoteDataSource: RemoteDataSourceItems
-
-    @Component.Factory
-    interface FactoryTest {
-        fun create(): ItemTestComponent
-    }
-}
 
 @Module
 class TestAppModule {
@@ -61,16 +43,6 @@ class TestAppModule {
     @Singleton
     fun championsRemoteDataSourceProvider():
             ChampionsRemoteDataSource = FakeChampionsRemoteDataSource()
-
-    @Provides
-    @Singleton
-    fun itemsLocalDataSourceProvider():
-            LocalDataSourceItems = FakeItemsLocalDataSource()
-
-    @Provides
-    @Singleton
-    fun itemsRemoteDataSourceProvider():
-            RemoteDataSourceItems = FakeItemsRemoteDataSource()
 }
 
 val defaultFakeChampions = listOf(
@@ -79,14 +51,6 @@ val defaultFakeChampions = listOf(
     localChampion.copy(id = "id3"),
     localChampion.copy(id = "id4"),
     localChampion.copy(id = "id5")
-)
-
-val defaultFakeItems = listOf(
-    localItem.copy(id = 1),
-    localItem.copy(id = 2),
-    localItem.copy(id = 3),
-    localItem.copy(id = 4),
-    localItem.copy(id = 5)
 )
 
 class FakeChampionsLocalDataSource : ChampionsLocalDataSource {
@@ -109,26 +73,6 @@ class FakeChampionsLocalDataSource : ChampionsLocalDataSource {
             = champions.first { it.id == championId }
 }
 
-class FakeItemsLocalDataSource : LocalDataSourceItems {
-
-    var items = emptyList<Item>()
-
-    override suspend fun itemsExists(): Boolean = items.isEmpty()
-
-    override suspend fun saveItems(items: List<Item>) {
-        this.items = items
-    }
-
-    override suspend fun getItems(): List<Item> = items
-
-    override suspend fun updateItem(item: Item) {
-        items = items.filterNot { it.id == item.id } + item
-    }
-
-    override suspend fun findItemById(itemId: String): Item
-            = items.first { it.id.toString() == itemId }
-}
-
 class FakeChampionsRemoteDataSource : ChampionsRemoteDataSource {
 
     var champions = defaultFakeChampions
@@ -137,10 +81,3 @@ class FakeChampionsRemoteDataSource : ChampionsRemoteDataSource {
 
 }
 
-class FakeItemsRemoteDataSource : RemoteDataSourceItems {
-
-    var items = defaultFakeItems
-
-    override suspend fun getItems(region: String): List<Item> = items
-
-}
